getioncoloc affiche la liste de colocations du propriétaire

+++ src/main/java/fr/formation/afpa/service/IUtilisateurService.java
@@ -12,10 +12,10 @@ public interface IUtilisateurService {
 
 	void delete(AppUser u);
 
-	void deleteByUserId(Long long1);
+	void deleteByUserId(int long1);
 
 	AppUser findByLogin(String login);
 
-	AppUser findByUserId(Long userId);
+	AppUser findByUserId(int userId);
 
 }
